refactor: move a few functiosn to helpers.rs
